+++ src/FamilyHubs.SharedKernel.Razor/Health/HealthChecksBuilderExtensions.cs
@@ -1,5 +1,4 @@
-﻿using FamilyHubs.SharedKernel.Razor.FamilyHubsUi.Options;
-using FamilyHubs.SharedKernel.Razor.FamilyHubsUi.Options.HealthCheck;
+﻿using FamilyHubs.SharedKernel.Razor.FamilyHubsUi.Options.HealthCheck;
 using HealthChecks.UI.Client;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Diagnostics.HealthChecks;
